fix(auth-runtime): decouple mainserver provisioning error class
